+++ mobile/app/(auth)/index.jsx
@@ -1,9 +1,137 @@
-import { View, Text } from 'react-native'
+/**
+ * @file Login.jsx
+ * @description Authentication screen providing a user login form 
+ * with field handling, password visibility toggling, and layout adjustment.
+ */
+import {
+    View,
+    Text,
+    Image,
+    TextInput,
+    TouchableOpacity,
+    ActivityIndicator,
+    KeyboardAvoidingView,
+    Platform,
+    Alert,
+} from "react-native";
+import { Link } from "expo-router";
+import styles from "../../assets/styles/login.styles";
+import { useState } from "react";
+import { Ionicons } from "@expo/vector-icons";
+import COLORS from "../../constants/colors";
 
+/**
+ * Login Screen Component
+ * Renders the interface for authenticating existing users into the app.
+ */
 export default function Login() {
+    // 1. State Hooks
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const [showPassword, setShowPassword] = useState(false);
+    const [isLoading, setIsLoading] = useState(false);
+    
+    // 2. Event Handlers
+    const handleLogin = async () => {
+        const result = await login(email, password);
+        
+        if (!result.success) Alert.alert("Error", result.error);
+    };
+    
+    // 3. UI Rendering
     return (
-        <View>
-            <Text>Login Screen</Text>
+        <KeyboardAvoidingView
+            style={{ flex: 1 }}
+            behavior={Platform.OS === "ios" ? "padding" : "height"}
+        >
+        <View style={styles.container}>
+            {/* ILLUSTRATION */}
+            <View style={styles.topIllustration}>
+                <Image
+                    source={require("../../assets/images/i.png")}
+                    style={styles.illustrationImage}
+                    resizeMode="contain"
+                />
+            </View>
+            
+            <View style={styles.card}>
+                <View style={styles.formContainer}>
+                    {/* EMAIL */}
+                    <View style={styles.inputGroup}>
+                        <Text style={styles.label}>Email</Text>
+                        <View style={styles.inputContainer}>
+                            <Ionicons
+                            name="mail-outline"
+                            size={20}
+                            color={COLORS.primary}
+                            style={styles.inputIcon}
+                            />
+                            <TextInput
+                            style={styles.input}
+                            placeholder="Enter your email"
+                            placeholderTextColor={COLORS.placeholderText}
+                            value={email}
+                            onChangeText={setEmail}
+                            keyboardType="email-address"
+                            autoCapitalize="none"
+                            />
+                        </View>
+                    </View>
+                
+                {/* PASSWORD */}
+                <View style={styles.inputGroup}>
+                    <Text style={styles.label}>Password</Text>
+                    <View style={styles.inputContainer}>
+                        {/* LEFT ICON */}
+                        <Ionicons
+                        name="lock-closed-outline"
+                        size={20}
+                        color={COLORS.primary}
+                        style={styles.inputIcon}
+                        />
+                        {/* INPUT */}
+                        <TextInput
+                        style={styles.input}
+                        placeholder="Enter your password"
+                        placeholderTextColor={COLORS.placeholderText}
+                        value={password}
+                        onChangeText={setPassword}
+                        secureTextEntry={!showPassword}
+                        />
+                        
+                        <TouchableOpacity
+                        onPress={() => setShowPassword(!showPassword)}
+                        style={styles.eyeIcon}
+                        >
+                            <Ionicons
+                            name={showPassword ? "eye-outline" : "eye-off-outline"}
+                            size={20}
+                            color={COLORS.primary}
+                            />
+                        </TouchableOpacity>
+                    </View>
+                </View>
+                
+                <TouchableOpacity style={styles.button} onPress={handleLogin} disabled={isLoading}>
+                    {isLoading ? (
+                        <ActivityIndicator color="#fff" />
+                    ) : (
+                        <Text style={styles.buttonText}>Login</Text>
+                    )}
+                </TouchableOpacity>
+                
+                    {/* FOOTER */}
+                    <View style={styles.footer}>
+                        <Text style={styles.footerText}>Don't have an account?</Text>
+                        <Link href="/signup" asChild>
+                            <TouchableOpacity>
+                            <Text style={styles.link}>Sign Up</Text>
+                            </TouchableOpacity>
+                        </Link>
+                    </View>
+                </View>
+            </View>
         </View>
+    </KeyboardAvoidingView>
     );
 }